fixed some errors

+++ api/controllers/AdminController.js
@@ -57,25 +57,29 @@ module.exports = {
     //adding new peoducts
     PostAddProduct: async(req,res)=>{
         try {
-            //description validation
-            if(!validator.isLength(req.body.description,{min:10})){
-                return res.badRequest({error:messages.descLength})
-                //return res.redirect('/admin/add-product')
+            console.log(req.user.superUser)
+            if(req.user.superUser){
+                //description validation
+                if(!validator.isLength(req.body.description,{min:10})){
+                    return res.badRequest({error:messages.descLength})
+                    //return res.redirect('/admin/add-product')
+                }
+                //image url validation
+                if(!validator.isURL(req.body.imageUrl)){
+                    return res.badRequest({error:messages.imgUrl})
+                    //return res.redirect('/admin/add-product')
+                }
+                //creating new product
+                const shop = await Shop.create({
+                    title:req.body.title,
+                    imageUrl:req.body.imageUrl,
+                    price:req.body.price,
+                    description:req.body.description,
+                    category:req.body.category
+                }).fetch()
+                return res.ok({success:messages.addProduct})
             }
-            //image url validation
-            if(!validator.isURL(req.body.imageUrl)){
-                return res.badRequest({error:messages.imgUrl})
-                //return res.redirect('/admin/add-product')
-            }
-            //creating new product
-            const shop = await Shop.create({
-                title:req.body.title,
-                imageUrl:req.body.imageUrl,
-                price:req.body.price,
-                description:req.body.description,
-                category:req.body.category
-            }).fetch()
-            return res.ok({success:messages.addProduct})
+            return res.badRequest({ error: messages.notAuthorized });
         } catch (error) {
              return res.serverError({error:error.message})
         }
@@ -85,27 +89,31 @@ module.exports = {
     postEditProduct:async(req,res)=>{
         try {
             //description validation
-            if(!validator.isLength(req.body.description,{min:10})){
-                return res.badRequest(messages.descLength)
-                //return res.redirect(`/admin/edit-product/:${req.params.id}`)
+            if(req.user.superUser){
+
+                if(!validator.isLength(req.body.description,{min:10})){
+                    return res.badRequest(messages.descLength)
+                    //return res.redirect(`/admin/edit-product/:${req.params.id}`)
+                }
+                //image url validation
+                if(!validator.isURL(req.body.imageUrl)){
+                    return res.badRequest(messages.imgUrl)
+                    //return res.redirect('/admin/edit-product')
+                }
+                const product = await Shop.findOne({id:req.body.productId})
+                if(!product){
+                    return res.status(400).send({error:messages.noProduct})
+                }
+                await Shop.update({id:product.id},{
+                    title:req.body.title,
+                    imageUrl:req.body.imageUrl,
+                    description:req.body.description,
+                    price:req.body.price,
+                    category:req.body.category
+                })
+                return res.ok({success:messages.editProduct})
             }
-            //image url validation
-            if(!validator.isURL(req.body.imageUrl)){
-                return res.badRequest(messages.imgUrl)
-                //return res.redirect('/admin/edit-product')
-            }
-            const product = await Shop.findOne({id:req.body.productId})
-            if(!product){
-                return res.status(400).send({error:messages.noProduct})
-            }
-            await Shop.update({id:product.id},{
-                title:req.body.title,
-                imageUrl:req.body.imageUrl,
-                description:req.body.description,
-                price:req.body.price,
-                category:req.body.category
-            })
-            return res.ok({success:messages.editProduct})
+            return res.badRequest({ error: messages.notAuthorized });
         } catch (error) {
              return res.serverError({error:error.message})
         }
@@ -114,16 +122,19 @@ module.exports = {
     //deleting proucts
     PostDeleteProduct: async(req,res)=>{
         try {
-            console.log('hdbc')
-            const product = await Shop.findOne({id:req.body.productId})
-            console.log(req.body.productId)
-            if(!product){
-                return res.status(ResCode.notFound).send({error:messages.noProduct})
+            if(req.user.superUser){
+                console.log('hdbc')
+                const product = await Shop.findOne({id:req.body.productId})
+                console.log(req.body.productId)
+                if(!product){
+                    return res.status(ResCode.notFound).send({error:messages.noProduct})
+                }
+                await Shop.destroy({id:product.id})
+                //console.log('after:',product)
+                await CartItem.destroy({shop:product.id})
+                return res.ok(messages.deleteProduct)
             }
-            await Shop.destroy({id:product.id})
-            //console.log('after:',product)
-            await CartItem.destroy({shop:product.id})
-            return res.ok(messages.deleteProduct)
+            return res.badRequest({ error: messages.notAuthorized });
         } catch (error) {
              return res.serverError({error:error.message})
         }
